Split out the test file
